+++ hahomematic/central_unit.py
@@ -688,7 +688,6 @@ class CentralUnit:
                     await self.remove_device(device=device)
             except KeyError:
                 _LOGGER.warning("delete_devices failed: Unable to delete: %s", address)
-        await self.paramset_descriptions.save()
 
     @callback_system_event(HH_EVENT_NEW_DEVICES)
     async def add_new_devices(