+++ changelog.md
@@ -6,6 +6,8 @@
 - Remove LOWBAT from HM-LC-Sw1-Pl, HM-LC-Sw2-FM
 - Simplify entity de-/registration
 - Refactor add/delete device and add tests
+- Add un_ignore_list to test config
+- Allow unignore for DEVICE_ERROR_EVENTS
 
 # Version 2023.1.0 (2023-01-01)
 - API Cleanup